+++ lib/wheneverd/schedule.rb
@@ -8,9 +8,14 @@ module Wheneverd
     # @return [Array<Entry>]
     attr_reader :entries
 
+    # @return [Array<Wheneverd::Service>]
+    attr_reader :services
+
     # @param entries [Array<Entry>]
-    def initialize(entries: [])
+    # @param services [Array<Wheneverd::Service>]
+    def initialize(entries: [], services: [])
       @entries = entries.dup
+      @services = services.dup
     end
 
     # Append an entry to the schedule.
@@ -19,6 +24,15 @@ module Wheneverd
     # @return [Schedule] self
     def add_entry(entry)
       entries << entry
+      self
+    end
+
+    # Append a long-running service to the schedule.
+    #
+    # @param service [Wheneverd::Service]
+    # @return [Schedule] self
+    def add_service(service)
+      services << service
       self
     end
   end